requirements.txt
torch==2.6.0
torchvision==0.21.0
numpy==1.26.4
Pillow>=9.0.0
scipy>=1.10.0
opencv-python>=4.7.0
matplotlib>=3.7.0
openai>=1.0.0
flask>=2.3.0
tqdm>=4.65.0
loguru>=0.7.0
trimesh>=3.20.0
transformers>=4.35.0,<5.0.0
pandas>=2.0.0
einops>=0.6.0
huggingface_hub
pyarrow